+++ src/Patch.php
@@ -6,13 +6,20 @@ use Attribute;
 use NckRtl\RouteMaker\Enums\HttpMethod;
 
 #[Attribute]
-class Route
+class Patch extends RouteAttribute
 {
     public function __construct(
-        public HttpMethod $method = HttpMethod::GET,
         public ?string $uri = null,
         public ?string $name = null,
         public ?array $parameters = null,
         public array|string|null $middleware = null,
-    ) {}
+    ) {
+        $this->method = HttpMethod::PATCH;
+        parent::__construct(
+            $uri,
+            $name,
+            $parameters,
+            $middleware,
+        );
+    }
 }