+++ tests/Fixtures/Controllers/Middleware/MiddlewareTestController.php
@@ -4,8 +4,9 @@ namespace NckRtl\RouteMaker\Tests\Http\Controllers\temp;
 
 use Illuminate\Routing\Controller;
 use Inertia\Response;
-use NckRtl\RouteMaker\Enums\HttpMethod;
-use NckRtl\RouteMaker\Route;
+use NckRtl\RouteMaker\Get;
+use NckRtl\RouteMaker\Post;
+use NckRtl\RouteMaker\Put;
 
 class MiddlewareTestController extends Controller
 {
@@ -13,25 +14,25 @@ class MiddlewareTestController extends Controller
 
     protected static array $routeMiddleware = ['controller-mw'];
 
-    #[Route(method: HttpMethod::POST, middleware: 'method-mw')]
+    #[Post(middleware: 'method-mw')]
     public function store(): Response
     {
         return inertia('Middleware/Store');
     }
 
-    #[Route(method: HttpMethod::GET, middleware: ['method-mw', 'another-mw'])]
+    #[Get(middleware: ['method-mw', 'another-mw'])]
     public function index(): Response
     {
         return inertia('Middleware/Index');
     }
 
-    #[Route(method: HttpMethod::GET, uri: 'no-method-mw')]
+    #[Get(uri: 'no-method-mw')]
     public function noMethodMiddleware(): Response
     {
         return inertia('Middleware/NoMethod');
     }
 
-    #[Route(method: HttpMethod::PUT, middleware: ['controller-mw', 'unique-method-mw'])]
+    #[Put(middleware: ['controller-mw', 'unique-method-mw'])]
     public function update(): Response
     {
         return inertia('Middleware/Update');